Introduce Policy

+++ app/Http/Controllers/PostController.php
@@ -31,9 +31,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        if (!$post->ownedBy(auth()->user())) {
+        $this->authorize('destroy', $post);
 
-        }
         $post->delete();
 
         return back();
